yawns in cry

+++ Team7_LibraryApp/Data/BookRankReport.cs
@@ -12,9 +12,9 @@ namespace Team7_LibraryApp.Data
         public string Title { get; }
         public string BookAuthor { get; }
         public string Genre { get; }
-        public int Rank { get; }
+        public long Rank { get; }
 
-        public BookRankReport(string ISBN, string Title, string BookAuthor, string Genre, int Rank)
+        public BookRankReport(string ISBN, string Title, string BookAuthor, string Genre, long Rank)
         {
             this.ISBN = ISBN;
             this.Title = Title;
